requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
altgraph==0.17.3
asgiref==3.6.0
auto-py-to-exe==2.27.0
backports.zoneinfo==0.2.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Eel==0.14.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
idna==3.4
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
